Enable API consumer to remove referees from matches
This could have been implemented as an extension from PUT to the same
resource, but DELETE is a more natural method for removing a resource.

A 404 is sent both if the match is non-existent, and if the match does
not have a referee. Although it may seem ambiguous it would not be a
clear API if anything other was sent if either was missing. This means
that API consumers may not be able to differentiate the two cases
programmatically, unless they try to parse the messages, which would not
be advised. Instead a good response from the consumer may be to simply
reload their information about referees to see if any action is needed.