Take first result from DB when creating competition
JDBC will return a list of results from DB queries. When creating a
competition the list will only contain one result. Returning this result
directly is what would be expected from the function.

This was the cause for the bug that caused the first admin creation to
fail as it relied on the return value of the first call to
`create-competition!`.