Mark all migrations as complete on init
New initializations of the database should not require any of the
current migrations to be ran, as the `init.sql` should be up-to-date
with all the changes from the migration scripts. Marking these as
completed will stop any of them to be executed the during the next
launch.

This allows the migration scripts to have reduced logic regarding when
they should be executed.